Add read method to Ecg class
Improved the Ecg class read method to be (slightly) more general, to
allow it to read the PTB-XL data as well as the Lobachevsky data from
physionet.org. A couple of other minor changes as well, including
introducing the start of a plot method to the class, and the option to
reset the time indices for individual beats when separating the signal.