Separate out get_peaks from get_beats
Separated out the process for getting the peaks of the signal from the
overall process of sorting into individual beats (potentially uselessly,
considering other bits in the pipeline!).

Also added an assert check on the read_ecg_from_igb to check the value
of dt being passed, making sure it's in ms not s.